// DesktopApp/src/types/SessionHistory.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { HistoryBlock } from "./HistoryBlock";
import type { SessionMeta } from "./SessionMeta";

export type SessionHistory = { meta: SessionMeta, blocks: Array<HistoryBlock>, hasMoreBefore: boolean, };